refactor: move walletPassModel to src/data to fix Vite middleware 500 error on Windows
